+++ lib/graphql/fragment_cache/object_helpers.rb
@@ -10,34 +10,32 @@ module GraphQL
     module ObjectHelpers
       extend Forwardable
 
-      NO_OBJECT = Object.new
+      def self.included(base)
+        return if base < GraphQL::Execution::Interpreter::HandlesRawValue
 
-      def_delegator :field, :connection?
+        base.include(GraphQL::Execution::Interpreter::HandlesRawValue)
+      end
+
+      NO_OBJECT = Object.new
 
       def cache_fragment(object_to_cache = NO_OBJECT, **options, &block)
         raise ArgumentError, "Block or argument must be provided" unless block_given? || object_to_cache != NO_OBJECT
 
         options[:object] = object_to_cache if object_to_cache != NO_OBJECT
 
-        fragment = Fragment.new(context, options)
+        context_to_use = options.delete(:context)
+        context_to_use = context if context_to_use.nil? && respond_to?(:context)
+        raise ArgumentError, "cannot find context, please pass it explicitly" unless context_to_use
+
+        fragment = Fragment.new(context_to_use, options)
 
         if (cached = fragment.read)
           return cached == Fragment::NIL_IN_CACHE ? nil : raw_value(cached)
         end
 
         (block_given? ? block.call : object_to_cache).tap do |resolved_value|
-          context.fragments << fragment
+          context_to_use.fragments << fragment
         end
-      end
-
-      private
-
-      def field
-        interpreter_context[:current_field]
-      end
-
-      def interpreter_context
-        @interpreter_context ||= context.namespace(:interpreter)
       end
     end
   end